hop/requirements.txt
fastapi==0.81.0
pydantic==1.10.1
uvicorn==0.18.3
discord.py==2.0.1
requests==2.28.1